rewritten code for mac compatibility

+++ aw_qt/__main__.py
@@ -1,42 +1,40 @@
-from aw_qt.main import main
 import os
 import sys
 import ctypes
+from aw_qt.main import main
+def show_error_message():
+    ctypes.windll.user32.MessageBoxW(0, "Sundial is already running.", "Sundial.exe", 0x10)
 
 def check_lock_file(lock_file_path):
     if os.path.exists(lock_file_path):
-        ctypes.windll.user32.MessageBoxW(0, 
-                                         "Sundial is already running.", 
-                                         "Sundial.exe", 
-                                         0x10)  # 0x10 is the code for an error icon
+        show_error_message()
         return True
     return False
 
-def create_lock_file(lock_file_path):
+def manage_lock_file(lock_file_path, create=False):
     try:
-        with open(lock_file_path, 'w') as lock_file:
-            lock_file.write("")
+        if create:
+            with open(lock_file_path, 'w') as lock_file:
+                lock_file.write("")
+        else:
+            os.remove(lock_file_path)
     except Exception as e:
         print(e)
         return False
     return True
 
-def delete_lock_file(lock_file_path):
-    try:
-        os.remove(lock_file_path)
-    except Exception as e:
-        print(e)
 
 
-lock_file_path = os.path.join(os.getenv('TEMP'), 'sundail.lock')
-
-if check_lock_file(lock_file_path):
-    sys.exit(0)
-
-if create_lock_file(lock_file_path):
-    try:
+if sys.platform == "win32":
+    lock_file_path = os.path.join(os.getenv('TEMP'), 'sundial.lock')
+    if check_lock_file(lock_file_path):
+        sys.exit(0)
+    if manage_lock_file(lock_file_path, create=True):
+        try:
+            main()
+        finally:
+            manage_lock_file(lock_file_path)
+elif sys.platform == "darwin":
         main()
-    finally:
-        delete_lock_file(lock_file_path)
